Added getArray functions.
It is possible to read arrays from a JSON file.
These functions assume that the values inside a array are from the
same type. This is the correct behaviour.

Although is much more simpler this way.
These functions returns std::vector with the correct type.
In my optinion this approach is more usable in real problems.